feat: Add lightweight prototyping framework
Add modular prototyping framework for rapid rendering technique development:

New modules:
- Application: PrototypeApplication, Camera, Scene (JSON-based), SceneObject,
  MousePicker (ray-AABB picking), GizmoController (ImGuizmo wrapper),
  ShaderWatcher (hot reload)
- Profiler: CPU/GPU timing with ImPlot visualization

Third-party integrations:
- ImPlot: Graph/chart visualization for profiler
- ImGuizmo: 3D gizmo for object manipulation

Key features:
- JSON-based scene loading with resource callbacks
- Ray-AABB intersection for mouse picking
- Shader hot reload via file watching
- CPU/GPU profiling with history graphs
- Gizmo controls (translate/rotate/scale) with snap support

Also includes:
- API macros (HS_PROFILER_API, HS_APPLICATION_API) for DLL export/import
- Sample scene files and configuration
- Design documentation (Docs/PrototypingFramework.md)

+++ Source/Engine/Renderer/RenderPath.h
@@ -42,9 +42,9 @@ public:
     private:
         RenderPath* _renderer;
 
-        std::unordered_map<uint32, RHIRenderPass*> _renderPassCache;
-        std::unordered_map<uint32, RHIFramebuffer*> _framebufferCache;
-        std::unordered_map<uint32, RHIGraphicsPipeline*> _gPipelineCache;
+        std::unordered_map<size_t, RHIRenderPass*> _renderPassCache;
+        std::unordered_map<size_t, RHIFramebuffer*> _framebufferCache;
+        std::unordered_map<size_t, RHIGraphicsPipeline*> _gPipelineCache;
     };
 
     RenderPath(RHIContext* rhiContext);
